+++ extensions/weather/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Feels Like] - 2023-08-25
 
 - Add `Temperature Feels Like` to command and menu
+- Add `UV Index` to command and menu
 
 ## [Add MenuBar] - 2022-11-26